Revert "Merge pull request #891 from YasinHossain/perf/surah-initial-cache"
This reverts commit 96ffbb139a776a8d8686797a73851cc96010f62e, reversing
changes made to 61212237c8e1a509938acf1bf35604d0d8bef541.
