+++ lib/api/server/index.ts
@@ -6,4 +6,3 @@
 export { getChaptersServer, getChapterServer } from './chapters';
 export { getVersesOfDayServer } from './verses';
 export { getTafsirVersePageDataServer } from './tafsir';
-export { getSurahInitialDataServer } from './surah';
